Add with-gpl-data feature for embedding MIME DB
Since the `builtin/magic`, `builtin/aliases` and `builtin/subclasses`
database files originate from the FreeDesktop project, they're the only
components of this library that prevent its use under a permissive
licence.

Here we change the default behaviour to search the standard XDG paths
for the MIME information database, and load it at runtime.

Included is a new `with-gpl-data` feature which conditionally compiles
the GPLed database into the library, the way it previously was by
default. The database itself is carved out into a separate workspace
crate, such that the main crate licence can be set back to MIT.