Fix some Clippy lints
Almost everything here is a syntax only change, the exception is the use
of `bytecount::count()` for searching for NULLs in `text/plain`.

In the current implementation, bytecount just uses a slightly more
efficient algorithm for finding the number of bytes, but optionally we
can enable a `runtime-dispatch-simd` feature and enable SSE2 and AVX2
codepaths if it would be valuable.