Refactor Checker definitions to use Traits
Rather than represent Checker as a big bag of function pointers in a
struct, we can use a Trait and &dyn references to those trait objects.

As we can store the &dyn reference directly in the _SUPPORT structure we
even reduce the amount of pointer chasing we do by 1, as the check
doesn't need to look up the index again.